+++ src/pages/404.ts
@@ -1,3 +1,3 @@
-import Page404View from "@views/404";
+import Page404View from "@views/404/404View";
 
 export default Page404View;
